LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A5.1_T1-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Parse error
S7.9_A5.1_T1

PASS Expected parsing failure

TEST COMPLETE
